Move test classes to mirror src structure
